+++ crossbeam-skiplist/benches/skiplist.rs
@@ -14,7 +14,7 @@ fn insert(b: &mut Bencher) {
     b.iter(|| {
         let map = SkipList::new(epoch::default_collector().clone());
 
-        let mut num = 0 as u64;
+        let mut num = 0u64;
         for _ in 0..1_000 {
             num = num.wrapping_mul(17).wrapping_add(255);
             map.insert(num, !num, guard);
@@ -27,7 +27,7 @@ fn iter(b: &mut Bencher) {
     let guard = &epoch::pin();
     let map = SkipList::new(epoch::default_collector().clone());
 
-    let mut num = 0 as u64;
+    let mut num = 0u64;
     for _ in 0..1_000 {
         num = num.wrapping_mul(17).wrapping_add(255);
         map.insert(num, !num, guard).release(guard);
@@ -45,7 +45,7 @@ fn rev_iter(b: &mut Bencher) {
     let guard = &epoch::pin();
     let map = SkipList::new(epoch::default_collector().clone());
 
-    let mut num = 0 as u64;
+    let mut num = 0u64;
     for _ in 0..1_000 {
         num = num.wrapping_mul(17).wrapping_add(255);
         map.insert(num, !num, guard).release(guard);
@@ -63,14 +63,14 @@ fn lookup(b: &mut Bencher) {
     let guard = &epoch::pin();
     let map = SkipList::new(epoch::default_collector().clone());
 
-    let mut num = 0 as u64;
+    let mut num = 0u64;
     for _ in 0..1_000 {
         num = num.wrapping_mul(17).wrapping_add(255);
         map.insert(num, !num, guard).release(guard);
     }
 
     b.iter(|| {
-        let mut num = 0 as u64;
+        let mut num = 0u64;
         for _ in 0..1_000 {
             num = num.wrapping_mul(17).wrapping_add(255);
             black_box(map.get(&num, guard));
@@ -85,13 +85,13 @@ fn insert_remove(b: &mut Bencher) {
     b.iter(|| {
         let map = SkipList::new(epoch::default_collector().clone());
 
-        let mut num = 0 as u64;
+        let mut num = 0u64;
         for _ in 0..1_000 {
             num = num.wrapping_mul(17).wrapping_add(255);
             map.insert(num, !num, guard).release(guard);
         }
 
-        let mut num = 0 as u64;
+        let mut num = 0u64;
         for _ in 0..1_000 {
             num = num.wrapping_mul(17).wrapping_add(255);
             black_box(map.remove(&num, guard).unwrap().release(guard));
